requirements.txt
requests==2.27.1
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==1.1.5
selenium==4.0.0a7
fake-useragent==1.4.0  # 兼容Python 3.8
retry==0.9.2
openpyxl==3.1.2
python-dotenv==0.20.0
flask==2.0.3
pymongo==4.1.1
APScheduler==3.10.4
certifi==2023.7.22
chardet==5.0.0
